Initial work on contact map flattening
This initial work will serve as the keyword generator for Mongo which
will be used as part of the search API. The keywords are just all of the
values that a given map contains in the form of a set which can be
easily traversed given some input.

This function is in place due to Mongo's lack of full-text search
support.